requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
Pillow==10.0.1
pydantic==2.3.0
pydantic_core==2.6.3
typing_extensions==4.7.1
urllib3==2.0.5
yarl==1.9.2